# Optimize LSTMPeephole by pre-calculating the input2hidden linear