## What changes were proposed in this pull request?
add preTopology in LSTMPeephole to boost the performance

## How was this patch tested?
Unit Test.

The structure of LSTMPeephole in BigDL and Torch are different. So I rearrange the weights in copy operation.


The speed:

I set batchSize = 12, inputSize = 300, hiddenSize = 300, and time = 200.
Running on my Desktop.
```
elapsed time | before    | after
------------------------------------------
1                | 1.25213 | 1.11945
2                | 1.25913 | 1.13069
3                | 1.27919 | 1.22337
4                | 1.27735 | 1.18582
5                | 1.23374 | 1.16419
------------------------------------------
average        | 1.26031 | 1.16470
------------------------------------------
throughput    | 9.5214   | 10.303 
```
Thus, the revised version will give a ~8.2% throughput improvement.
With larger batchSize, inputSize and hiddenSize, the improvement will be more salient.